safari-534.53-branch/LayoutTests/platform/qt/http/tests/security/xss-DENIED-xsl-external-entity-redirect-expected.txt

CONSOLE MESSAGE: line 3: <html xmlns='http://www.w3.org/1999/xhtml/'><body><p>Running an XSL-T 1.0 stylesheet with a 2.0 processor.</p></body></html>
This test includes a cross-origin external entity. It passes if the load fails and thus there is no text below this line.
